@@ -13,7 +13,6 @@
     "lint:fix": "biome lint --fix",
     "lint:ci": "biome ci"
   },
-  "comment": "IMPORTANT! date-fns is pinned to 1.3.x due to https://github.com/mui-org/material-ui-pickers/issues/1440",
   "dependencies": {
     "@auth0/auth0-react": "^2.3.0",
     "@babel/runtime": "^7.28.2",
@@ -25,12 +24,13 @@
     "@codemirror/state": "^6.5.2",
     "@codemirror/theme-one-dark": "^6.1.3",
     "@codemirror/view": "^6.38.1",
-    "@date-io/date-fns": "^1.3.13",
-    "@material-ui/core": "^4.11.2",
-    "@material-ui/icons": "^4.11.2",
-    "@material-ui/lab": "4.0.0-alpha.57",
-    "@material-ui/pickers": "^3.3.10",
-    "@material-ui/styles": "^4.11.2",
+    "@emotion/react": "^11.14.0",
+    "@emotion/styled": "^11.14.1",
+    "@mui/icons-material": "^7.3.1",
+    "@mui/lab": "^7.0.0-beta.16",
+    "@mui/material": "^7.3.1",
+    "@mui/system": "^7.3.1",
+    "@mui/x-date-pickers": "^8.10.0",
     "@tanstack/react-table": "^8.21.3",
     "@tanstack/react-virtual": "^3.13.12",
     "@uiw/codemirror-theme-material": "^4.25.1",
@@ -39,19 +39,19 @@
     "change-case": "^5.4.4",
     "classnames": "^2.2.6",
     "codemirror": "^6.0.2",
-    "date-fns": "2.26.0",
+    "date-fns": "^4.1.0",
     "deep-sort-object": "^1.0.2",
     "diff": "^8.0.2",
     "mdi-react": "^8.2.0",
     "prop-types": "^15",
     "qs": "^6.10.1",
     "ramda": "^0.31.3",
-    "react": "^17",
+    "react": "^19",
     "react-diff-view": "^3.3.2",
-    "react-dom": "^17",
-    "react-helmet": "^6.1.0",
+    "react-dom": "^19",
     "react-number-format": "^5.4.4",
-    "react-router-dom": "^5.2.0"
+    "react-router-dom": "7.8.0",
+    "tss-react": "^4.9.19"
   },
   "devDependencies": {
     "@babel/core": "^7.28.0",
@@ -68,7 +68,6 @@
     "jest": "^30.0.5",
     "jest-environment-jsdom": "30.0.5",
     "lodash.omit": "^4.5.0",
-    "mdi-react": "^4.0.0",
     "mini-css-extract-plugin": "^2.9.2",
     "react-refresh": "^0.17.0",
     "webpack": "^5.100.2",
