@@ -1,32 +1,51 @@
 import { Auth0Provider } from '@auth0/auth0-react';
-import CssBaseline from '@material-ui/core/CssBaseline';
-import { ThemeProvider } from '@material-ui/styles';
-import React, { Fragment } from 'react';
+import CssBaseline from '@mui/material/CssBaseline';
+import { StyledEngineProvider, ThemeProvider } from '@mui/material/styles';
+import React from 'react';
+import { GlobalStyles } from 'tss-react';
 import Main from './Main';
 import theme from './theme';
 
 const App = () => {
   return (
-    <Fragment>
-      <CssBaseline />
-      <Auth0Provider
-        domain={process.env.AUTH0_DOMAIN}
-        clientId={process.env.AUTH0_CLIENT_ID}
-        redirectUri={process.env.AUTH0_REDIRECT_URI}
-        audience={process.env.AUTH0_AUDIENCE}
-        scope={process.env.AUTH0_SCOPE}
-        authorizationParams={{
-          audience: process.env.AUTH0_AUDIENCE,
-          scope: process.env.AUTH0_SCOPE,
-        }}
-        leeway={30}
-        cacheLocation="localstorage"
-      >
-        <ThemeProvider theme={theme}>
+    <StyledEngineProvider injectFirst>
+      <ThemeProvider theme={theme}>
+        <GlobalStyles
+          styles={{
+            'html, body': {
+              height: '100%',
+            },
+            '#root': {
+              height: '100%',
+            },
+            '.cm-editor': {
+              fontSize: 13,
+              height: '100%',
+            },
+            '.cm-theme': {
+              flex: 1,
+              overflow: 'auto',
+            },
+          }}
+        />
+        <CssBaseline />
+        <Auth0Provider
+          domain={process.env.AUTH0_DOMAIN}
+          clientId={process.env.AUTH0_CLIENT_ID}
+          redirectUri={process.env.AUTH0_REDIRECT_URI}
+          audience={process.env.AUTH0_AUDIENCE}
+          scope={process.env.AUTH0_SCOPE}
+          authorizationParams={{
+            audience: process.env.AUTH0_AUDIENCE,
+            scope: process.env.AUTH0_SCOPE,
+          }}
+          leeway={30}
+          cacheLocation="localstorage"
+        >
           <Main />
-        </ThemeProvider>
-      </Auth0Provider>
-    </Fragment>
+        </Auth0Provider>
+      </ThemeProvider>
+    </StyledEngineProvider>
   );
 };
 
