@@ -1,20 +1,20 @@
-import MuiSpeedDial from '@material-ui/lab/SpeedDial';
-import SpeedDialIcon from '@material-ui/lab/SpeedDialIcon';
-import { makeStyles } from '@material-ui/styles';
+import MuiSpeedDial from '@mui/material/SpeedDial';
+import SpeedDialIcon from '@mui/material/SpeedDialIcon';
 import classNames from 'classnames';
 import CloseIcon from 'mdi-react/CloseIcon';
 import DotsVerticalIcon from 'mdi-react/DotsVerticalIcon';
 import { arrayOf, node, object, oneOfType } from 'prop-types';
 import React, { useState } from 'react';
+import { makeStyles } from 'tss-react/mui';
 
-const useStyles = makeStyles((theme) => ({
+const useStyles = makeStyles()((theme) => ({
   speedDial: {
     ...theme.mixins.fab,
   },
 }));
 
 function SpeedDial({ children, className, FabProps, ...props }) {
-  const classes = useStyles();
+  const { classes } = useStyles();
   const [open, setOpen] = useState(false);
   const handleClick = () => setOpen(!open);
   const handleClose = () => setOpen(false);
