@@ -1,18 +1,19 @@
-import Card from '@material-ui/core/Card';
-import CardContent from '@material-ui/core/CardContent';
-import List from '@material-ui/core/List';
-import ListItem from '@material-ui/core/ListItem';
-import ListItemText from '@material-ui/core/ListItemText';
-import Typography from '@material-ui/core/Typography';
-import { makeStyles } from '@material-ui/styles';
+import Card from '@mui/material/Card';
+import CardContent from '@mui/material/CardContent';
+import List from '@mui/material/List';
+import ListItem from '@mui/material/ListItem';
+import ListItemButton from '@mui/material/ListItemButton';
+import ListItemText from '@mui/material/ListItemText';
+import Typography from '@mui/material/Typography';
 import LinkIcon from 'mdi-react/LinkIcon';
 import React from 'react';
+import { makeStyles } from 'tss-react/mui';
 import Dashboard from '../../components/Dashboard';
 import balrogSrc from '../../images/balrog.svg';
 import { RULES_COMMON_FILTERS } from '../../utils/constants';
 import Link from '../../utils/Link';
 
-const useStyles = makeStyles((theme) => ({
+const useStyles = makeStyles()((theme) => ({
   balrogImage: {
     width: 800,
     height: 800,
@@ -30,7 +31,7 @@ const useStyles = makeStyles((theme) => ({
 }));
 
 function Home() {
-  const classes = useStyles();
+  const { classes } = useStyles();
 
   return (
     <Dashboard title="Home">
@@ -42,12 +43,16 @@ function Home() {
           </Typography>
           <List>
             {RULES_COMMON_FILTERS.map(({ link, label }) => (
-              <Link className={classes.link} key={label} to={link}>
-                <ListItem button>
+              <ListItem key={label} disablePadding>
+                <ListItemButton
+                  component={Link}
+                  className={classes.link}
+                  to={link}
+                >
                   <ListItemText primary={label} />
                   <LinkIcon />
-                </ListItem>
-              </Link>
+                </ListItemButton>
+              </ListItem>
             ))}
           </List>
         </CardContent>
