@@ -3,10 +3,13 @@ import { matchPath } from 'react-router-dom';
 // Returns an array of matched routes.
 const matchRoutes = (path, routes, branch = []) => {
   const matchingRoute = routes.find((route) =>
-    matchPath(path, {
-      path: route.path,
-      exact: route.exact,
-    }),
+    matchPath(
+      {
+        path: route.path,
+        end: route.exact,
+      },
+      path,
+    ),
   );
 
   if (matchingRoute) {
